requirements.txt
langchain>=0.3.24,<0.4.0
langchain-openai>=0.3.8,<0.4.0
langchain-chroma>=0.2.2,<0.3.0
langchain-text-splitters>=0.3.8,<0.4.0
chromadb>=0.5.23,<0.7.0
openai>=1.68.0,<2.0.0
SQLAlchemy>=2.0.29,<3.0.0
streamlit>=1.43.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.10.6,<3.0.0
